(DOC-842) Confusing language around a code sample
This patch rewords the language around the code sample in the "On the
Command Line" section of the "Enable and Disable Puppet Agent" section
of the PE 3.x Orchestration documentation. It also adds a bit of
markup and the word "Example" to make it more consistent with other
examples on this page.