Problem:
The documentation uses `/etc/puppet/hieradata/` in some places and
`/etc/puppet/hiera/` in others. We must choose one or the other.

As a result, many of the examples on this page do not work. To test,
simply run through the examples in order and note that hiera is unable
to find anything including the keys in common.yaml:

```
[root@puppet3 puppet]# cat hiera.yaml

---
:backends:
  - yaml
:yaml:
  :datadir: /etc/puppet/hieradata
:hierarchy:
  - "nodes/%{::fqdn}"
  - "virtual/%{::virtual}"
  - "osfamily/%{osfamily}"
  - common
[root@puppet3 puppet]# ls hiera
common.yaml  nodes  osfamily  virtual
[root@puppet3 puppet]# head hiera/common.yaml
vmwaretools::version: 8.6.5-621624
[root@puppet3 puppet]# service puppetmaster restart
Stopping puppetmaster:                                     [  OK  ]
Starting puppetmaster:                                     [  OK  ]
[root@puppet3 puppet]# hiera vmwaretools::version
nil
[root@puppet3 puppet]#
```

Solution:

Standardize on `/etc/puppet/hieradata/`, since that seems to be the most
common path in the other PuppetDocs documentation. Replace
`/etc/puppet/hiera/` with `/etc/puppet/hieradata`.