(DOCUMENT-126) Qualify variable as used in hierarchy
When hiera recieves values from the command line, they are just the
literal name that it will look up internally. The :: doesn't mean
anything special to hiera. What this means is that when a variable is
referenced as ::foo, then the value provided on the command line needs
to use ::foo.

There may have been a hiera version in the past that allowed the
documented version to work, but it doesn't do that anymore. :(