(DOC-61) Version Custom Facts guide, update links
Facter 2.x has had its own version of Custom Facts guide, but the old
version of it lived on in the /guides/ directory. This commit splits
that out into 1.6 and 1.7 versions and updates a slew of links so that
they point to the right places. I've already sent a PR to redirect
the /guides/ version to /facter/latest/, but in the meantime I'll leave
it up.