(DOCUMENT-129) Add warning notes about PassengerHighPerformance
We talked to Ryan McKern about this, since he'd recently audited our vhost
configs for performance. He pointed out that the setting can be changed in vhost
or directory scope, which means you can disable it for only certain routes.

I added notes about what it interferes with to the Passenger and multi-master
pages, and a quick example on how to disable it. I also moved the example
declaration to within the puppet master vhost, so at least it wouldn't clobber
other modules globally.