Fix missing URL on a learning puppet page
Earlier this year, I merged an old update that got lost in the run-up to PE 3.0.
Since it was written before the 3.0 docs were posted, I didn't have a URL to
point to for class params in the console, so I used a placeholder instead. I
didn't notice that when I merged it.