(guts) Upgrade jquery to version 1.8.3
We can't upgrade beyond version 1.9 because it made breaking changes, including
the removal of some deprecated methods that our three or four jquery scripts all
seemed to rely on. However, going to 1.8.3 gets us access to all the improved
methods that replaced the deprecated ones, so we have a starting point for
revising them.