(guts) One-character fix for busted drop-down menus
The (outdated, stylistically out-of-step) drop-down menus in the site header
broke when I updated jquery to 1.8.3. It turns out this is because the hasClass
method used to accept a . at the start of the class name, the way you'd write it
in a selector, and somewhere between then and now they decided it had to just be
the NAME of a class.