(guts) Death to that hacky {% iflink %} tag; normal links now work fine
I added a re-do of the pe 3.3 sidebar to show what I mean here.

The one sketchy case would be where you have absolute links to a version and are
looking at the doc via the "/latest" shortcut. Basically, the way puppetdb does
it. Once we overhaul the way nav snippets are maintained, we can get around
that; in the meantime, either puppetdb can keep using the {% iflink %} tags or
we can live with having nav sections collapsed on the first page you view of the
/latest/ docs.