DOCS 1140: edited backslash requirement in Windows package provider `source` attribute
Added info to resources_packages_windows for 3.7 that forward- and backslashes are accepted by the `source` attribute for Windows package provider

Added info to resources_package_windows for 3.6 that only backslashes are acceptable in `source`, as this information was missing from the page.

Changed lang_windows_file_paths for 3.6 to correctly reflect the `source` backslash requirement.