Fix (most?) mixed/insecure content warnings
This should hopefully change the protocol on all images and javascripts we use
in the site to HTTPS. It also changes a bunch of links.

There's one element where we're not sure what will happen, since it looks like
the certificate on the HTTPS site might be from a non-standard CA or something.
We're looking into it. (http://info.puppetlabs.com/js/mktFormSupport.js)