Don't imply ordering promises about regex node definitions
We were saying Puppet would use the "first" matching regex node if there was
more than one match, which is mostly true, modulo some weirdness if a regex
stringifies into an exact hostname when you remove the special characters. But
in practice, which one wins is irrelevant, because you shouldn't be overlapping
your node regexes in the first place! So don't do that.