Guts: Fix broken nav sidebars when viewing "latest" in puppet, puppet-server, puppetdb
The nav javascript compares (a slightly munged version of) the portion of the
URL after the server to the link destination when checking whether a link points
to the current page. This worked great in sidebars that use relative links,
because regardless whether you were in, e.g. the 3.7 docs or the "latest" docs,
the browser would calculate the path correctly before trying to compare it to
the current page. This DIDN'T work so well with absolute links, and the puppet,
puppet-server, and puppetdb docs all use those now for various annoying but
temporarily important reasons.

This commit changes the logic in that JS so that a path segment of "latest" will
also match any continuous string of digits and periods. That should be good
enough to get by for now.