(guts) Enable adding a version note to each page, and to directories of pages
To add version notes to a directory full of pages, use the version_notes key in
source/_config.yml. The most specific path for a given page will determine which
note it gets.

To add a version note to a single page, use the version_note key in its yaml
frontmatter.

Version notes are fragments of arbitrary HTML. Optionally, you can use
`<span class="noisy">` for notes that should pop a bit more.

This commit has example items in _config.yml. We'll add more permanent items in
a future commit.