(maint) Update future_lang_namespaces autoload documentation
This updates the autoload documentation to correctly describe how it
maps a qualified name to a file to load. This is important since users
needs to know who the search is done to find what could otherwise be
mysterious bugs, or to enable them to read code that does not follow the
style guide "one object per file" advice.

The change also includes a description of how nested namespace work
(when defining classes and defines inside classes). While not
recommended, this information is needed to be able to read code that
does not follow the style guide.