(maint) Update future_lang_relationships with 4x features
There are fewer restrictions in 4x language than in 3x. This adds
information about what this means for creating relationships betwen
resources.

It is possible to place resource creation in arrays or hashes, or as
esults of blocks. This gives the user the ability to use the produced
references to resources after they have been created. This commit adds
several examples how this can be used.