Guts: Add new "partial" tag -- {% partial ./_repeated_text.md %}
This new tag works just like "include," except it can insert text fragments from
anywhere in the source directory. You can use relative paths or absolute ones.
("Absolute" paths actually start at the source directory, not at the root of the
filesystem.)

Also, note that Jekyll will "hide" files whose names start with an underscore,
and not transfer them to the final output; this is a good way to make sure your
snippets are available in the source, but don't end up exposed on the site.