Lift the requirement for aligned hashrockets
Hashrockets have been required to be aligned for a long time. This
has long been attributed to readability. I wonder if there is real
data to support this, or if readability is a matter of preference.

At OpenStack, we find that the aligned hashrockets restriction causes
unnecessary churn in the code base. Every time a new parameter is added
with a different length, the entire code bock shifts in or out.

There are two major problems from this. Removing the restriction removes
the problem:

1) git-blame doesn't show correct information. Instead of showing the
last person/commit that materially changed the line in question, it shows
the last person/commit that tabbed the line in or out.
There are whitespace-ignoring options that can be used on the command
line, but there is a lot of tooling around git and not all of it runs
in this mode.

2) Code review has more code changing. What should be a one line change
to review now looks like several other parameters have been changed or
added. At OpenStack we review a lot of code, increasing the speed that
we can review at is the single biggest way to move our project forward
faster.

This came out of a discussion this review:
https://review.openstack.org/#/c/164819/