(DOCUMENT-368) alternate ways to get 64 bit paths
On Windows, the search path is able to be used to search for a process. This allows
us to take advantage of using the search path to enter a specific order of searches
for an executable so that we are always using the 64-bit version or alternatively,
always selecting the 32-bit version based on where the executable is found first on
the path. This allows a process that is sometimes in sysnative and sometimes in
system32 based on whether we are in a 32-bit process on a 64-bit OS or in a 32-bit
process on a 32-bit OS. Either way the proper item will be selected provided the
search strings are correct.