(guts) Add ability to set toc: columns in YAML frontmatter
This will do something a lot like what the revamped type reference does, with
several columns of text instead of a huge scroll of doom. Best for TOCs that
include a whole load of one-word items, worst for TOCs that include lots of
nested levels and a lot of sentence-length headers.