(DOCUMENT-434) Handle package type's multiple namevars in type reference
Right now, there's only one core resource type that uses a compound uniqueness
key. Instead of jumping straight to a general-purpose fix, let's do a special
case fix for that one and see how long it holds up. Due to the way we generate
the affected reference, this is a simple template change.