Move PC1 docs to includes and versioned docs.
We refer to Puppet Collections (PC) frequently throughout the
installation and upgrade documentation, and Puppet Collections are
designed to be version-specific bundles of packages. However, our PC
documentation is in a legacy unversioned guide covering package
repositories.

*   To version the documentation, add a page to the versioned
    documentation describing Puppet Collections.
*   Since much of this information is also useful in step-by-step
    installation and upgrade guides, implement as much of the PC content
    as Jekyll includes so we can maintain common content across PCs.
*   Update the installation and upgrade guides with relevant links and
    Jekyll includes.

Once we've confirmed this all works as expected, we can port these
changes back to previous Puppet documentation versions, remove the
unversioned guide, and redirect it or repair links to it.