Guts: Make TOC plugin not barf on ID-less headers
The TOC plugin was doing something naïve that could result in calling [] on
nil. It was also not being very smart about attribute strings that include
anything more than just an ID; if you added a class to a header, it would give up
on finding the ID. Now it should be a lot more resilient.