(glitchy) Disable pygments highlighter
At some point in the history of Jekyll between 0.11.2 and 2.5.3, they changed
the behavior of the {% highlight ruby %} tag when Pygments is disabled, and now
it outputs HTML compatible with prism.js, which is what we use for our fenced
code blocks.

That means that if we disable Pygments and remove it from our toolchain, our few
remaining Liquid highlight tags will render the same way as the rest of our code
blocks.

Also, something about the way Jekyll uses Pygments now blows up when you give it
a language it doesn't recognize... or something. Since this solution was ready
to hand, I didn't feel like investigating the problem the rest of the way.

The remaining problem as of this commit is the sitemap, which gets deleted.