(guts) Add json gem to Gemfile
I don't know how this has been working on my machine, but on a CentOS 7 box
running Ruby 2.0.0, Jekyll blew up because it couldn't load Ruby's JSON library.
Perhaps it's rolled into core in OS X's Ruby build? Or perhaps it's pulled in as
a dependency by something in the :generate_references gem group, which we
exclude on the builders! Whichever: we'd better add it to the list so it works
on the machines where it matters most.