DOC-1455 (REFERENCEPALOOZA): Prune and relocate the generated references
This commit moves references for the newest .z version of each x.y series into
/puppet/VERSION/reference, and deletes the remaining generated references.

For 3.8 through 4.3, I also went ahead and regenerated the refs, for two
reasons: First, we (I) had gotten lazy about generating new dot-z references, so
most of them were slightly out of date. Second, the improved reference generator
makes a page-per-resource type version of the type reference, and this ensures
we have that available for all of our currently maintained version series (as
well as the now unmaintained versions in-between).

This commit DOESN'T deal with redirects, changing existing links, or updating
the navigation snippets.