(guts) Convince Bundler to always add the local lib directory to the load path
The `gemspec` directive (usually used when developing a gem) makes Bundler
always add the local `lib` directory to the load path. This is great for us,
because it lets us put general utility code in the puppet_docs and
puppet_references namespaces and easily load it from things like the Rakefile,
Jekyll plugins, etc. without having to munge the load path everywhere.

As part of convincing Bundler to do the thing, I also had to add a
puppet_docs.gemspec file. It doesn't have any other purpose, since we're not
releasing this helper code as a gem, but I made it look plausible anyway.