(DOC-2491) Change the format of externalsources to use the URL as the key
Previously, the key used in the externalsources hash was entirely redundant. We
were using it as the (hidden, total implementation detail) filename of the
working copy, and that was it. Well, instead, how about we use the destination
URL as the uniqueness key? Less redundant typing, and we can just munge it to
make a predictably boring directory name, same as we do for repo URLs.