(guts) Intercept Jekyll's config file and modify it before building site
We're already using the PuppetDocs::Config class to munge the config file a bit
before using it in the Rakefile, and I think it makes sense to do the same thing
for any modifications of config data for Jekyll's use. So we'll write (and delete)
a temporary amended config file whenever we run a Jekyll command.