(guts) Avoid conflict in templates with site.documents method
The Jekyll site object has a "documents" method, and when Jekyll merges the info
in the site.config hash into the top level of site (when preparing to render
Liquid templates), it won't include our very important "documents" hash.

Well, that's okay; we're munging this data before Jekyll gets it anyway. Add a
duplicate under the non-conflicting name "document_list".