(guts) Fix `rake serve`'s busted index.html behavior
There was some middleware in config.ru meant to mimic Apache's handling of
index.html. Unfortunately, it could only catch URLs ending with a slash
(/pe/3.8/), so if you omitted it (/pe/3.8) you'd get the directory listing
instead of the index.html you were OBVIOUSLY trying to reach (I mean COME ON).

This commit replaces the middleware with a class that inherits from
Rack::Directory, so we can go ahead and stat the index.html file when deciding
whether to serve a directory listing. So now the local webserver shouldn't
💩 on perfectly normal URLs.