(DOC-2507) Use doc data to build canonical link in the absence of explicit one
This should result in less redundant nonsense to maintain.

HOWEVER, it's a bit limited at the moment: it uses the "real" URL of latest,
instead of the /latest/ shortcut. I think that ought to be fine, having the
canonical one change every few months. And we'll find out whether google
disagrees.