(guts) Set shortcut variables for linking to documents, like {{puppet}}, etc.
With this commit, you can link to the "locally correct" version of a given
document with something like...

[classes]: {{puppet}}/lang_classes.html

If you're linking from a versioned document that knows about its own version of
the destination document (via _config.yml), this will resolve to the correct
version. If not, it falls back to the latest version.