Remove redundant version note from foss SSL regen guide
This is the only page where we ever used the ability to set a single-page
version note, and we only used it to duplicate the big warning that's already at
the top of the page.

The real warning is noisy enough, and the red version note is subject to
banner-blindness anyway. Remove it.