(guts) Use {% partial %} and frontmatter defaults to replace {% render_nav %}
Now that the {% partial %} tag can accept variables, and Jekyll lets you set
default variables for pages based on path prefix (albeit using a data format
that's unwieldy to work with in YAML), the render_nav tag is basically obsolete.
This commit replaces it with two extensions we've doubled down on (the "partial"
tag and PuppetDocs::Config), so there are fewer things to reason about.

This commit does not change the interface at all; nav snippets are still set via
the defaultnav config, the nav key in yaml frontmatter, or document data.