Manage redirects and improve deployments
This adds a redirect configuration file, so that redirects don't have to
be managed through pull requests on puppetlabs-modules.

Deployment process
------------------

This changes the deployment process to make it faster, more reliable,
and less likely to be noticed by users. An outline of the new steps:

 1. rsync assets to stage
 2. rsync stage to tarball source directory
 3. create the tarball
 4. move the tarball into stage
 5. swap stage and final (site is now live)
 6. check the NGINX rewrite configuration file
 7. reload NGINX if the configuration file is valid

This process reduces the amount of time during which the live site is
changing. Swapping the two directories is nearly instantaneous, but
rsyncing takes quite a while.

This also means that the tarball will be readily as soon as the new site
is made live.

Directory structure
-------------------

This changes the directory structure of the deployment target. There is
now a containing directory, e.g. /var/www/docs.puppetlabs.com, with
final and stage subdirectories, and an nginx_rewrite.conf file.

The staging directory for the tarball is now also in the containing
directory, instead of ~docsdeploy.

No longer syncing modification times
------------------------------------

For background, Jekyll updates the modification time of pretty much
every file when it builds. Rsync uses those modifications times to
determine what files have changed, so it thinks every file has changed.

This switches to using fast checksums in rsync, which ignores the
modification times, and doesn't update them on the server when they
change on the machine the deployment is running from.

The primary benefit is that deployments run more than 5 times faster.

A side benefit of this is that timestamp based caching will work better
for users. It's hard for me to tell how useful that would be, but
roughly 2% of requests over the last 6 hours on the production web site
were "304 Not Modified", which indicates caching was in use.