(guts) Fix display of in-page TOCs with more than two header levels
You have to keep track of all the header levels you've descended into if you
don't want to end up with weird glitches.

By the way, this code is some of the more difficult to maintain stuff I've ever
written. If there's a smart way to make a linear list of headers into a
hierarchy of nested lists, we could really use it. I'll check with some of our
real programmers and see if they know an algorithm for that.