Revise home page.
The heights of the two sections at the center of the docs home page
don't always match. The CSS already implements these divs as CSS tables,
but because the borders are built on divs nested inside the cells
instead of the cells themselves, the borders don't take advantage of
this configuration to automatically align when the box heights differ.

This commit revises the home page HTML and CSS to apply the borders to
the table-cell divs instead of the inner boxes, and adds border-spacing
to recreate the gap between boxes. As a result, the site-width, header,
and sub-footer divs also must be adjusted to accomodate the different
alignment produced by the use of border-spacing instead of padding.

Also, wrap the sub-footer in a similar box as the columns, and create
four evenly spaced columns instead of a horizontal list of links to
accommodate future subfooter expansion.