(guts) Link checking: Fix false positives for "/hiera/latest" et al.
We were registering broken paths for URLs that use the "latest" symlink but
don't follow it with "/anything", even though they're generally working links
that go to the front page of the document. This adds a special case to resolve
"latest" if it's the last component in a path.

This commit also stops trimming the trailing slash from paths like
"/hiera/latest/" (although that had no bad consequences, since we check for
trimmed slashes later).