[DEV] Added elements and assets for new brand to the docs site files
Added the fonts Calibre and Inconsolatat to the files>fonts folder, added the
social-icons, puppet-logo, puppet-dags to the images folder.

Altered the style.css doc to remove unneeded styles and the styles that are being replaced,
mostly around the header/footer. There was a lot of duplication and over specificity that
I cleaned up too while I was in there. Added a TOC to the css doc.

Changed the index,html and default.html files to have the new header and footer elements.