(guts) Simplify the document switcher: only go to "best" version
Previously, we presented up to two versions for each document: the latest
version, and the "best" version for the document you're currently in. The idea
was to allow easy switching between related document versions (e.g. PE 15.3,
Puppet 4.3, and Server 2.2), while also supporting a "reset" if you were just in
the wrong section of the site entirely.

On further consideration, that's too complicated, so we'll just support
switching to the "best" version, followed by using the version switcher if
that's not what you want. This should be less confusing even for users who do
want to reset to latest.

This commit also leverages the {{puppet}} style of magic URLs, which we
implemented after we built the original version of the document switcher.