(guts) Fix {{puppet}} within Puppet docs (e.g.)
Magic URLs like {{puppet}} are supposed to resolve to the base URL of the
document you're linking to. (So, something like /puppet/4.3/reference, for the
above.) But there was a corner case if you were _already in_ the document you
were linking to, and it would resolve to the version number instead of the base
URL. This commit fixes that.