(javascript) End the vile scourge of anchor-jump
When clicking the doc switchers, it always jumps you a centimeter or two to the
top of the page if you aren't already scrolled all the way up. This is because
the main element is an `<a href="#">` link.

But it turns out to be really easy to make it not do that! Hallelujah.