(guts) Make link checker skip link-like text in code block examples
An example XML snippet included the line "src='tools\**'", and the link checker
was registering it as a broken link. It should really ignore those, and we can
do that by just making sure we only look for src and href attributes that are
_inside_ an html tag. Since code blocks with XML and other markup have their
angle brackets escaped as &lt; and &gt;, this should keep them out of the mix.