(guts) Add rudimentary broken link reporting script (link_report.rb)
This relies on the `rake build_and_check_links` task, which creates a YAML file.
All the script does is filter that YAML file and pretty-print the info you want.

The code is kinda crap because it doesn't need to not be crap. :) We can improve
this later as needed.