(css) Fix numbers in ordered list items; add support for nested ones
In our previous CSS resets, we weren't resetting the list style for ordered
lists and ended up using the browser default... which was fine. But now we have
better resets, so we have to specify the style explicitly.

While I'm in here doing that, I might as well make nested ordered lists behave
like people expect from, e.g. MS Word and the like. So now the second level of
list will be a., b., etc., and the third level will be i., ii., etc.