(css) Fix numbering on in-page TOCs
These shouldn't have numbers; they're OL items due to historical wrongness.
Unfortunately, now that we're setting a list style for OLs in the body section,
that style is technically "more specific" than the one for the TOC according to
the style cascade rules. So we'll just put an !important on it to jump ahead in
the specificity line.