Add col-xs-12 to nav sidebar (to fix links)
Without this, the sidebar nav div was (sort of, kind of) taking up the whole
vertical space of the enclosing container, which made it intercept all clicks on
links in the main content div (and probably doing some other wacky stuff). By
clarifying that we want it to take up the whole width on mobile, it stays in its
vertical lane. ¯\_(ツ)_/¯