Remove Marketo forms style tweaks
This changes the spacing of elements on /downloads/index.html, but in a way
I don't consider detrimental or even particularly noticeable. TBH I think it
improves it. By the way, this form brings in styling of its own via the javascript
that inserts it into the page -- if the authors believed it should be styled in
a particular way, that's the way they'd style it.