Add "Jump to table of contents" link on sm and xs viewports
On smaller viewports, the sidebar nav gets pushed down below the page content,
and it can take a maddening amount of scrolling to reach it. This should solve
95% of that problem in a nice, non-overkill way.

I think grouping this link with the document/version switchers should help
show which TOC we're talking about (the document-wide one, not the in-page one).