Delete and redirect inventory service and PDL pages
For the 3.7 and 3.8 deprecated features pages, I did something a bit special by
linking to the inventory_service.markdown file in a commit on Github where it
still existed. This is because those pages are specifically about the removal of
a feature, so it remains useful to talk about the feature itself instead of its
replacement (but not useful enough to keep it posted on our live website).