Guts: Update prism.js to commit b9a9de5 (fix regexp detection in puppet highlighter)
The regexp detection in the Puppet plugin was screwing up bad on strings like
`"${master_config_dir}/.ssh"`, so I fixed it up a bit. I think this commit is
not yet in a released version of Prism, but it's been merged into the dev
branch, so it should be in the next version we update to.