DOC-2859: Use Puppet lexer (not ruby) to highlight Puppet code (puppet 4.5)
Traditionally, we didn't have the option of a Puppet lexer, and went with Ruby
because it yielded less-worse results. But it has glitches, like weird handling
of the :: namespace separator. So now that we have a purpose-built lexer, we
should always use it.

This commit includes a lot of hand-checking to make sure we don't accidentally
mis-highlight any actual ruby code.