Guts/references: Fetch repo upon instantiation of helper object
The version table generators had a mysterious behavior where they'd miss the
latest version the first time you ran them, but build up-to-date tables on the
next run. The problem was that my git repo helper object was delaying its fetch
until the first time I did a checkout, and I wasn't doing any checkouts until I
had already gotten the list of tags to work with.

There was really no point to delaying the fetch, so I pushed it into the init
method. Now we should have an up-to-date set of tags before we go looking for
tags.