Modernise usage of facts in language references
Prior to this commit, the example code around conditionals and the section on facts did not make use of structured facts.

This commit updates a couple of sections to show examples with modern structured-fact equivalents of old single value facts. It also describes how to access deeply nested fact values using hash syntax.

(Potentially controversial: I've opted to not quote hash keys, because doing so on 'puppet apply' leads to backslash-itis, and I find it visually noisy. AFAIK it is not semantically necessary but this might be making a unilateral style guide call that's not 100% OK)