Clarify Defined Resource Type Example
The current example for defined resource types is unclear, due to the
coding style used.

In general, listing all defined type parameters on one line is
considered poor style and very difficult to read.

In addition, comments in-line in a resource, between parameters, are
also viewed as poor style, and make the resource harder to read,
especially for new users who are just learning.

Hopefully this refactor clarifies the example.

Even as an "expert" Puppet user, I found the original fairly hard to
read.