Make project landing pages self-maintaining
Updating these lists was pointless bullshit, and since literally everything else
we do was more important, we kept forgetting to do it. Anyway, since we already
HAVE all the info about which versions exist, which ones are in which PE, which
one is the latest, which ones are pre-releases, etc. etc., um, WHY DON'T WE JUST
USE IT??? Exactly. Please enjoy never updating these again.