(guts) Make version tables use cached data, if present
This should speed up the version table generator, if you already have some
cached PE or agent JSON data in the references_output/version_tables folder. It
does introduce the possibility of stale or incorrect data if there's bad stuff
in the cache, so delete the version_tables dir if you run into that and it
should fix it right up.