(guts) Switch to a locally hosted copy of jquery
We've been using the "googleapis" servers for jquery because that's what
Bootstrap defaulted us to. But we ship a tarball of the docs for offline use,
and if your browser throws away its cache of the googleapis URL, it breaks our
sidebar and version switcher and stuff!

So let's switch to a local copy of that library. This should let us provide
equivalent experience offline, except for the marketing/tracking code.