(guts) Stop doing busted cowboy version math in version tables
Time to re-learn this lesson AGAIN.

At some point, I might go take out all the begin/rescue blocks. I'm trying to
protect against the chance of passing Versionomy something like `master` (which
makes it blow up), but now that I think about it, IDK if that's actually a
possibility, given the code paths that leave here. But tonight isn't the time to
actually read through all those and make sure of it. tl;dr begin/rescue is
probably the right thing to do in that sort_descending method, but it's
probably wrong for the stuff in version_tables/util.