(guts) Delete garbage directories that block external source symlinks
I've seen this a few times, and it goes a little like this:

* Your checkout of puppet-docs is old enough that the PE 3.7 or 3.8 docs used to
  be local instead of in a separate repo.
* At some point, you edited those soon-to-be-moved PE docs with a merge tool
  that left a .orig file behind, or changed Finder's view settings while
  innocently viewing that folder, or did any number of other things that could
  leave behind a (possibly invisible) file that Git is configured to ignore
  instead of automatically deleting.
* This leaves a source/pe/3.8 directory in puppet-docs, even though Git thinks
  there's no content there.
* When you build the site, our tools try to symlink the
  externalsources/<PE 3.8 checkout>/source directory into place at
  source/pe/3.8. HOWEVER, since that's a directory, the system misinterprets
  that command to mean "symlink it into place at source/pe/3.8/source." Which is
  NOT where it belongs!
* Later during the build, the template tries to find the nav snippet for that PE
  version at source/pe/3.8/_pe_toc.html, but the file isn't there, it's at
  source/pe/3.8/source/_pe_toc.html instead. The build freaks out and blows up,
  with a mysterious error message.

THIS SHOULD FIX THAT.

I'm making the assumption that any directory path "claimed" by an external
source (via the _config.yml file, as usual) should never exist, and if it DOES
exist, it's full of garbage and can be safely deleted. There's a small chance of
that deleting real data, but you'd have to go so far off-road to make that
happen that I'm going to not worry about it.