Adopt the frozen-in-amber Mco 1.2 docs as static content
Backstory: We added this via an external source ages ago, to support a CS
request for antique Mco docs for a probably now EOL-ed PE version. But the
external source was on MY personal github account, not the Puppet Labs one, and
we should definitely no longer be doing that. So this commit moves it all INTO
the docs repo, because there's no good reason for that to be an external source
anyway.