(PDK-463) Improve PowerShell 2 and 3 documentation
This commit updates the explanation for how to install the PDK on
systems with PowerShell v2 or v3. It corrects the assumption that you
can load the PowerShell by name and replaces it with the full path. It
also adds steps to load the module on v2 and v3 in your PowerShell
profile so it can be done automatically like v4 and v5 can.