(DOC-3292) Fix broken Puppet Server links in TOCs.
In the Puppet 5.0 and 5.1 TOCs, the link to Scaling Puppet Server
pointed to a previous page name. Fix this link.

The Puppet 5.2 TOC appears to have been copied from an earlier version
of Puppet that did not include recent additions to the Puppet Server
documentation. Fix the Puppet Server section of the TOC to make it
consistent with the 5.0 and 5.1 TOCs.