(DOC-3369) Remove or change references to Puppet 4
The "Updating 3.x manifests" page was written for Puppet 4 and not
changed when Puppet 5 was released, presumably because there were no
additional relevant changes. However, the heading and text of the page
suggest that the page is not relevant to Puppet 5.

Remove and edit references to Puppet 4 in this document to make it more
version-neutral.