(DOC-3411) Remove resource_type man page and links.
resource_type was removed in Puppet 5.0 by PUP-6123, but we still link
to it from the navigation for Puppet 5.0 onward, and published a man
page for it in Puppet 5.2 onward.

Remove the resource_type man page link from the sidebar navigation in
all Puppet 5 versions.

Remove the resource_type man page from the Puppet 5.2 and 5.3
documentation.