Line edit on strings docs
Line 34: Do you have info about WHY they would want to set YARD options for Strings? How do they know if they want to do this?

Line 56, which is step 2 in generating HTML docs: This should be a two-option step: a) To generate the documentation for the entire module, run `puppet strings` b) To generate the documentation for specific files or directories in a module ... and the rest of lines 60-72.

Also, I added "and view" to the title of this topic, and so you can make the two "view" instructions into steps in the task as well (lines 74 and 76).

Line 118: "a Rake task" -- can the user create their own Rake task, of which strings:gh_pages:update is an example? If so, then start line 120 with "For example, ". If you mean that they should be using strings:gh_pages:update, say "generate and publish HTML documentation with the strings:gh_pages:update Rake tasks, available in ..." (that is, merge lines 118 and 120 into one statement).

Line 187, line 191 - I think the "Markdown,JSON" and "If not specified..." entries are left over from a copy-paste and should be empty.

Line 190 makes no sense to me right now, but maybe that's because I haven't read the style guide. But maybe some context about what docstring text is? I dunno.

Lines 201-202 - mention "above" but those things aren't "above". I suspect the link will also be broken?

Line 214 feels like it could use an example

Lin 217: "Should"? If it _must_ be 140 characters or fewer, say that.