copy edit on strings style guide
Line 51: I took out the "140 chars" guidance because you've already given it as a general guideline, and it seems to be a soft requirement. Readability guidance needn't be repeated as though it were a hard-and-fast rule. If I'm misunderstanding the purpose of the 140 char guideline, it's because of Line 39, which should be fixed if it's not correct.

(I added some whitespace so my line numbers from hereon refer to the new line numbering)

Line 66: Add ", indented with two spaces" to the end of the second sentence?

Line 82: Move "This is an example of how to document a Puppet class." out of the comment and into regular text (or just delete). Same with Line 232.

Line 294 TODO question: Marking them in bold makes sense to me. Double-asterisks should work in the Markdown, yeah? We will have to experiment with the DITA, but <strong> tags should work in codeblocks.